Refactor AWS key generators to a separate package
Allows using them from other packages.
